# [Merged by Bors] - perf(topology/sheaves/presheaf_of_functions): speedups