Speed up`CommRing_yoneda` (which caused [a bors failure](https://github.com/leanprover-community/mathlib/actions/runs/3211416019/jobs/5249570763)) from >20s to <2.5s by filling in automatic fields. (merged in another PR)

Speed up `presheaf_to_Types` from 13.3s to 0.3s and `presheaf_to_Type` from 4s to 0.1s by filling in automatic fields.

---
[Zulip](https://leanprover.zulipchat.com/#narrow/stream/144837-PR-reviews/topic/16721.20coatoms.20in.20a.20t1_space)

<!-- The text above the `---` will become the commit message when your
PR is merged. Please leave a blank newline before the `---`, otherwise
GitHub will format the text above it as a title.

To indicate co-authors, include lines at the bottom of the commit message 
(that is, before the `---`) using the following format:

Co-authored-by: Author Name <author@email.com>

Any other comments you want to keep out of the PR commit should go
below the `---`, and placed outside this HTML comment, or else they
will be invisible to reviewers.

If this PR depends on other PRs, please list them below this comment,
using the following format:
- [ ] depends on: #abc [optional extra text]
- [ ] depends on: #xyz [optional extra text]
-->

[![Open in Gitpod](https://gitpod.io/button/open-in-gitpod.svg)](https://gitpod.io/from-referrer/)