@alreadydone 2022-10-09
Since this has been merged into #15663 which is in queue, let me take the opportunity of this PR to speedup to other declarations.