@kbuzzard 2022-10-09
You could leave a comment saying that tidy can do these but you're doing them anyway, but it's probably better to get this merged ASAP