@alreadydone 2022-10-09
> You could leave a comment saying that tidy can do these but you're doing them anyway, but it's probably better to get this merged ASAP

I think in general it's reasonable to prefer longer, less-automatic but much faster code over shorter but slower code, and does not need a special explanation. And it's clearly not rare to fill in the `map_id'/comp'` fields manually as can be seen from [the search results](https://github.com/leanprover-community/mathlib/search?p=1&q=map_id).

By the way, I just found out you actually beat me in opening PR (#16872) and had exactly the same changes!